# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mahayojana API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Mahayojana Backend",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "API Support",
      "email": "support@mahayojana.com"
    }
  },
  "servers": [
    {
      "url": "/v1",
      "description": "API V1"
    }
  ],
  "components": {
    "schemas": {
...
      "Policy": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "policyNumber": {
            "type": "string",
            "description": "Unique policy number"
          },
          "policyType": {
            "type": "string",
            "description": "Type of policy"
          },
          "title": {
            "type": "string",
            "description": "Policy title"
          },
          "description": {
            "type": "string",
            "description": "Policy description"
          },
          "policyStartDate": {
            "type": "string",
            "format": "date-time",
            "description": "Policy start date"
          },
          "policyEndDate": {
            "type": "string",
            "format": "date-time",
            "description": "Policy end date"
          },
          "policyStatus": {
            "type": "string",
            "description": "Current status of the policy"
          },
          "policyAmount": {
            "type": "number",
            "description": "Policy amount"
          },
          "policyTerm": {
            "type": "number",
            "description": "Policy term in months"
          },
          "schemaDefinition": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DynamicFieldDefinition"
            },
            "description": "Schema definition for dynamic fields"
          },
          "dynamicFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "Dynamic fields data based on schema definition"
          }
        }
      },
...
      "CreatePolicyRequest": {
        "type": "object",
        "required": [
          "policyNumber",
          "policyType",
          "title",
          "policyStartDate",
          "policyEndDate",
          "policyStatus",
          "policyAmount",
          "policyTerm"
        ],
        "properties": {
          "policyNumber": {
            "type": "string",
            "description": "Unique policy number"
          },
          "policyType": {
            "type": "string",
            "description": "Type of policy"
          },
          "title": {
            "type": "string",
            "description": "Policy title"
          },
          "description": {
            "type": "string",
            "description": "Policy description"
          },
          "policyStartDate": {
            "type": "string",
            "format": "date-time",
            "description": "Policy start date"
          },
          "policyEndDate": {
            "type": "string",
            "format": "date-time",
            "description": "Policy end date"
          },
          "policyStatus": {
            "type": "string",
            "description": "Current status of the policy"
          },
          "policyAmount": {
            "type": "number",
            "description": "Policy amount"
          },
          "policyTerm": {
            "type": "number",
            "description": "Policy term in months"
          },
          "schemaDefinition": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DynamicFieldDefinition"
            },
            "description": "Schema definition for dynamic fields"
          },
          "dynamicFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "Dynamic fields data based on schema definition"
          }
        }
      },
      "UpdateSchemaRequest": {
        "type": "object",
        "required": [
          "schemaDefinition"
        ],
        "properties": {
          "schemaDefinition": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DynamicFieldDefinition"
            },
            "description": "Schema definition for dynamic fields"
          }
        }
      },
      "UpdateDynamicFieldsRequest": {
        "type": "object",
        "required": [
          "dynamicFields"
        ],
        "properties": {
          "dynamicFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "Dynamic fields data based on schema definition"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/policies": {
      "post": {
        "summary": "Create a new policy",
        "tags": [
          "Policies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policy created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Policy created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all policies with pagination and filtering",
        "tags": [
          "Policies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "policyType",
            "schema": {
              "type": "string"
            },
            "description": "Filter by policy type"
          },
          {
            "in": "query",
            "name": "policyStatus",
            "schema": {
              "type": "string"
            },
            "description": "Filter by policy status"
          },
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "Filter by title (partial match)"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by start date (from)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by end date (to)"
          },
          {
            "in": "query",
            "name": "dynamic.*",
            "description": "Filter by dynamic fields (e.g., dynamic.category=OBC)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Policies retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "policies": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Policy"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/policies/{id}": {
      "get": {
        "summary": "Get a policy by ID",
        "tags": [
          "Policies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Policy ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Policy retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Policy not found"
          }
        }
      },
      "put": {
        "summary": "Update a policy",
        "tags": [
          "Policies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Policy ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "policyNumber": {
                    "type": "string"
                  },
                  "policyType": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "policyStartDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "policyEndDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "policyStatus": {
                    "type": "string"
                  },
                  "policyAmount": {
                    "type": "number"
                  },
                  "policyTerm": {
                    "type": "number"
                  },
                  "schemaDefinition": {
                    "type": "object",
                    "additionalProperties": {
                      "$ref": "#/components/schemas/DynamicFieldDefinition"
                    }
                  },
                  "dynamicFields": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policy updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Policy updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Policy not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a policy",
        "tags": [
          "Policies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Policy ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Policy deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Policy deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Policy not found"
          }
        }
      }
    },
    "/policies/{id}/schema": {
      "put": {
        "summary": "Update policy schema definition",
        "description": "Update the schema definition for dynamic fields of a policy",
        "tags": [
          "Policies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Policy ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchemaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schema updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Policy schema updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Policy not found"
          }
        }
      }
    },
    "/policies/{id}/fields": {
      "put": {
        "summary": "Update policy dynamic fields",
        "description": "Update the dynamic fields data for a policy",
        "tags": [
          "Policies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Policy ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDynamicFieldsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dynamic fields updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Policy dynamic fields updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Policy not found"
          }
        }
      }
    }
  },
...
}